+++ mjx/mujoco/mjx/_src/types.py
@@ -15,10 +15,8 @@
 """Base types used in MJX."""
 
 import enum
-from typing import List, Optional
 
 import jax
-import jax.numpy as jp
 import mujoco
 from mujoco.mjx._src.dataclasses import PyTreeNode  # pylint: disable=g-importing-member
 import numpy as np
@@ -213,6 +211,22 @@ class BiasType(enum.IntEnum):
   # unsupported: MUSCLE, USER
 
 
+class ConstraintType(enum.IntEnum):
+  """Type of constraint.
+
+  Attributes:
+    EQUALITY: equality constraint
+    LIMIT_JOINT: joint limit
+    CONTACT_PYRAMIDAL: frictional contact, pyramidal friction cone
+  """
+  EQUALITY = mujoco.mjtConstraint.mjCNSTR_EQUALITY
+  # unsupported: FRICTION_DOF, FRICTION_TENDON
+  LIMIT_JOINT = mujoco.mjtConstraint.mjCNSTR_LIMIT_JOINT
+  # unsupported: LIMIT_TENDON, CONTACT_FRICTIONLESS
+  CONTACT_PYRAMIDAL = mujoco.mjtConstraint.mjCNSTR_CONTACT_PYRAMIDAL
+  # unsupported: CONTACT_ELLIPTIC
+
+
 class CamLightType(enum.IntEnum):
   """Type of camera light.
 
@@ -371,6 +385,7 @@ class Model(PyTreeNode):
     geom_solref: constraint solver reference: contact         (ngeom, mjNREF)
     geom_solimp: constraint solver impedance: contact         (ngeom, mjNIMP)
     geom_size: geom-specific size parameters                  (ngeom, 3)
+    geom_rbound: radius of bounding sphere                    (ngeom,)
     geom_pos: local position offset rel. to body              (ngeom, 3)
     geom_quat: local orientation offset rel. to body          (ngeom, 4)
     geom_friction: friction for (slide, spin, roll)           (ngeom, 3)
@@ -388,17 +403,13 @@ class Model(PyTreeNode):
     cam_poscom0:  global position rel. to sub-com in qpos0    (ncam, 3)
     cam_pos0:  global position rel. to body in qpos0          (ncam, 3)
     cam_mat0:  global orientation in qpos0                    (ncam, 9)
-    mat_rgba: rgba                                            (nmat, 4)
-    mesh_vertadr: first vertex address                        (nmesh x 1)
-    mesh_faceadr: first face address                          (nmesh x 1)
+    mesh_vertadr: first vertex address                        (nmesh,)
+    mesh_faceadr: first face address                          (nmesh,)
+    mesh_graphadr: graph data address; -1: no graph           (nmesh,)
     mesh_vert: vertex positions for all meshes                (nmeshvert, 3)
     mesh_face: vertex face data                               (nmeshface, 3)
-    geom_convex_face: vertex face data, MJX only              (ngeom,)
-    geom_convex_vert: vertex data, MJX only                   (ngeom,)
-    geom_convex_edge_dir: unique edge direction, MJX only     (ngeom,)
-    geom_convex_facenormal: normal face data, MJX only        (ngeom,)
-    geom_convex_face_edge: edges for each face                (ngeom,)
-    geom_convex_face_edge_normal: edge normals for each face  (ngeom,)
+    mesh_graph: convex graph data                             (nmeshgraph,)
+    mat_rgba: rgba                                            (nmat, 4)
     pair_dim: contact dimensionality                          (npair,)
     pair_geom1: id of geom1                                   (npair,)
     pair_geom2: id of geom2                                   (npair,)
@@ -435,6 +446,7 @@ class Model(PyTreeNode):
     actuator_gear: scale length and transmitted force         (nu, 6)
     numeric_adr: address of field in numeric_data             (nnumeric,)
     numeric_data: array of all numeric fields                 (nnumericdata,)
+    name_meshadr: mesh name pointers                          (nmesh,)
     name_numericadr: numeric name pointers                    (nnumeric,)
     names: names of all objects, 0-terminated                 (nnames,)
   """
@@ -516,6 +528,7 @@ class Model(PyTreeNode):
   geom_solref: jax.Array
   geom_solimp: jax.Array
   geom_size: jax.Array
+  geom_rbound: jax.Array
   geom_pos: jax.Array
   geom_quat: jax.Array
   geom_friction: jax.Array
@@ -535,18 +548,14 @@ class Model(PyTreeNode):
   cam_mat0: jax.Array
   mesh_vertadr: np.ndarray
   mesh_faceadr: np.ndarray
+  mesh_graphadr: np.ndarray
   mesh_vert: np.ndarray
   mesh_face: np.ndarray
+  mesh_graph: np.ndarray
   mat_rgba: np.ndarray
   pair_dim: np.ndarray
   pair_geom1: np.ndarray
   pair_geom2: np.ndarray
-  geom_convex_face: List[Optional[jax.Array]]
-  geom_convex_vert: List[Optional[jax.Array]]
-  geom_convex_edge_dir: List[Optional[jax.Array]]
-  geom_convex_facenormal: List[Optional[jax.Array]]
-  geom_convex_edge: List[Optional[jax.Array]]
-  geom_convex_edge_face_normal: List[Optional[jax.Array]]
   pair_solref: jax.Array
   pair_solreffriction: jax.Array
   pair_solimp: jax.Array
@@ -580,6 +589,7 @@ class Model(PyTreeNode):
   actuator_gear: jax.Array
   numeric_adr: np.ndarray
   numeric_data: np.ndarray
+  name_meshadr: np.ndarray
   name_numericadr: np.ndarray
   names: bytes
 
@@ -596,8 +606,11 @@ class Contact(PyTreeNode):
     solref: constraint solver reference, normal direction             (mjNREF,)
     solreffriction: constraint solver reference, friction directions  (mjNREF,)
     solimp: constraint solver impedance                               (mjNIMP,)
-    geom1: id of geom 1
-    geom2: id of geom 2
+    dim: contact space dimensionality: 1, 3, 4, or 6
+    geom1: id of geom 1; deprecated, use geom[0]
+    geom2: id of geom 2; deprecated, use geom[1]
+    geom: geom ids                                                    (2,)
+    efc_address: address in efc; -1: not included
   """
   dist: jax.Array
   pos: jax.Array
@@ -607,33 +620,25 @@ class Contact(PyTreeNode):
   solref: jax.Array
   solreffriction: jax.Array
   solimp: jax.Array
-  # unsupported: mu, H, dim
+  # unsupported: mu, H
+  dim: np.ndarray
   geom1: jax.Array
   geom2: jax.Array
-  # unsupported: efc_address, exclude
-
-  @classmethod
-  def zero(cls, ncon: int = 0) -> 'Contact':
-    """Returns a contact filled with zeros."""
-    return Contact(
-        dist=jp.zeros(ncon),
-        pos=jp.zeros((ncon, 3,)),
-        frame=jp.zeros((ncon, 3, 3)),
-        includemargin=jp.zeros(ncon),
-        friction=jp.zeros((ncon, 5)),
-        solref=jp.zeros((ncon, mujoco.mjNREF)),
-        solreffriction=jp.zeros((ncon, mujoco.mjNREF)),
-        solimp=jp.zeros((ncon, mujoco.mjNIMP,)),
-        geom1=jp.zeros(ncon, dtype=int),
-        geom2=jp.zeros(ncon, dtype=int),
-    )
+  geom: jax.Array
+  # unsupported: flex, elem, vert, exclude
+  efc_address: np.ndarray
 
 
 class Data(PyTreeNode):
-  r"""Dynamic state that updates each step.\
+  r"""Dynamic state that updates each step.
 
   Attributes:
-    solver_niter: number of solver iterations, per island         (mjNISLAND,)
+    ne: number of equality constraints
+    nf: number of friction constraints
+    nl: number of limit constraints
+    nefc: number of constraints
+    ncon: number of contacts
+    solver_niter: number of solver iterations
     time: simulation time
     qpos: position                                                (nq,)
     qvel: velocity                                                (nv,)
@@ -663,14 +668,15 @@ class Data(PyTreeNode):
     cinert: com-based body inertia and mass                       (nbody, 10)
     actuator_length: actuator lengths                             (nu,)
     actuator_moment: actuator moments                             (nu, nv)
-    crb: com-based composite inertia and mass                     (nbody, 10)
+    crb: com-based composite inertia and mass                     (nbody, 10)  \
     qM: total inertia                                  if sparse: (nM,)
                                                        if dense:  (nv, nv)
     qLD: L'*D*L (or Cholesky) factorization of M.      if sparse: (nM,)
                                                        if dense:  (nv, nv)
     qLDiagInv: 1/diag(D)                               if sparse: (nv,)
                                                        if dense:  (0,)
-    contact: list of all detected contacts                        (ncon,)
+    contact: all detected contacts                                (ncon,)
+    efc_type: constraint type                                     (nefc,)
     efc_J: constraint Jacobian                                    (nefc, nv)
     efc_frictionloss: frictionloss (friction)                     (nefc,)
     efc_D: constraint mass                                        (nefc,)
@@ -689,6 +695,12 @@ class Data(PyTreeNode):
     efc_force: constraint force in constraint space               (nefc,)
     userdata: user data, not touched by engine                    (nuserdata,)
   """
+  # constant sizes:
+  ne: int
+  nf: int
+  nl: int
+  nefc: int
+  ncon: int
   # solver statistics:
   solver_niter: jax.Array
   # global properties:
@@ -732,6 +744,7 @@ class Data(PyTreeNode):
   qLD: jax.Array  # pylint:disable=invalid-name
   qLDiagInv: jax.Array  # pylint:disable=invalid-name
   contact: Contact
+  efc_type: np.ndarray
   efc_J: jax.Array  # pylint:disable=invalid-name
   efc_frictionloss: jax.Array
   efc_D: jax.Array  # pylint:disable=invalid-name