+++ doc/changelog.rst
@@ -14,12 +14,30 @@ General
 
 MJX
 ^^^
-3. Added cylinder plane collisions.
+.. admonition:: Breaking API changes
+   :class: attention
+
+   3. Removed deprecated ``mjx.device_get_into`` and ``mjx.device_put`` functions as they lack critical new
+      functionality.
+
+      **Migration:** Use ``mjx.get_data_into`` instead of ``mjx.device_get_into``, and ``mjx.put_data`` instead of
+      ``mjx.device_put``.
+
+4. Added cylinder plane collisions.
+5. Added ``efc_type`` to ``mjx.Data`` and ``dim``, ``efc_address`` to ``mjx.Contact``.
+6. Added ``geom`` to ``mjx.Contact`` and marked ``geom1``, ``geom2`` deprecated.
+7. Added ``ne``, ``nf``, ``nl``, ``nefc``, and ``ncon`` to ``mjx.Data`` to match ``mujoco.MjData``.
+8. Given the above added fields, removed ``mjx.get_params``, ``mjx.ncon``, and ``mjx.count_constraints``.
+9. Changed the way meshes are organized on device to speed up collision detection when a mesh is replicated for many
+   geoms.
+10. Fixed a bug where capsules might be ignored in broadphase colliision checking.
 
 Bug fixes
 ^^^^^^^^^
-4. Defaults of lights were not being saved, now fixed.
-5. Prevent overwriting of frame names by body names when saving an XML.  Bug introduced in 3.1.4.
+11. Defaults of lights were not being saved, now fixed.
+12. Prevent overwriting of frame names by body names when saving an XML.  Introduced in 3.1.4.
+13. Fixed bug in Python binding of :ref:`mj_saveModel`: ``buffer`` argument was documented as optional but was actually
+    not optional.
 
 
 Version 3.1.4 (April 10th, 2024)